test: move verification test into pd/src/verify.rs
